Add hand_players table, remove live props from table_seats, use unsigned ints

+++ database/migrations/2022_03_08_105920_create_cards.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->increments('id');
             $table->string('name');
             $table->string('abbreviation');
-            $table->integer('ranking');
+            $table->unsignedTinyInteger('ranking');
         });
 
         Schema::create('suits', function (Blueprint $table) {
