+++ database/migrations/2022_03_08_113123_create_hand_types_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('hand_types', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->integer('ranking');
+            $table->unsignedTinyInteger('ranking');
         });
     }
